+++ src/components/badge/format.ts
@@ -4,7 +4,6 @@ import NumAbbr from 'number-abbreviate';
  * Abbreviate a number
  *
  * @param {number} value the number to abbreviate
- *
  * @returns {string} abbreviated number
  */
 export function abbreviate(value: number): string {