+++ src/components/select/examples/select.tsx
@@ -12,16 +12,20 @@ export class SelectExample {
     @State()
     public disabled = false;
 
+    @State()
+    public readonly = false;
+
+    @State()
+    public required = false;
+
+    @State()
+    public invalid = false;
+
     private options: Option[] = [
         { text: 'Luke Skywalker', value: 'luke' },
         { text: 'Han Solo', value: 'han', disabled: true },
         { text: 'Leia Organo', value: 'leia' },
     ];
-
-    constructor() {
-        this.onChange = this.onChange.bind(this);
-        this.toggleEnabled = this.toggleEnabled.bind(this);
-    }
 
     public render() {
         return (
@@ -32,13 +36,32 @@ export class SelectExample {
                     value={this.value}
                     options={this.options}
                     disabled={this.disabled}
-                    onChange={this.onChange}
+                    readonly={this.readonly}
+                    required={this.required}
+                    invalid={this.invalid}
+                    onChange={this.changeHandler}
                 />
                 <p>
                     <limel-flex-container justify="end">
-                        <limel-button
-                            onClick={this.toggleEnabled}
-                            label={this.disabled ? 'Enable' : 'Disable'}
+                        <limel-checkbox
+                            checked={this.disabled}
+                            label="Disabled"
+                            onChange={this.setDisabled}
+                        />
+                        <limel-checkbox
+                            checked={this.readonly}
+                            label="Readonly"
+                            onChange={this.setReadonly}
+                        />
+                        <limel-checkbox
+                            checked={this.required}
+                            label="Required"
+                            onChange={this.setRequired}
+                        />
+                        <limel-checkbox
+                            checked={this.invalid}
+                            label="Invalid"
+                            onChange={this.setInvalid}
                         />
                     </limel-flex-container>
                 </p>
@@ -47,11 +70,27 @@ export class SelectExample {
         );
     }
 
-    private onChange(event) {
+    private changeHandler = (event) => {
         this.value = event.detail;
-    }
+    };
 
-    private toggleEnabled() {
-        this.disabled = !this.disabled;
-    }
+    private setDisabled = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.disabled = event.detail;
+    };
+
+    private setReadonly = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.readonly = event.detail;
+    };
+
+    private setRequired = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.required = event.detail;
+    };
+
+    private setInvalid = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.invalid = event.detail;
+    };
 }