+++ src/components/select/select.template.tsx
@@ -4,6 +4,7 @@ import { isMultiple } from '../../util/multiple';
 
 interface SelectTemplateProps {
     disabled?: boolean;
+    readonly?: boolean;
     required?: boolean;
     invalid?: boolean;
     options: Option[];
@@ -45,6 +46,7 @@ export const SelectTemplate: FunctionalComponent<SelectTemplateProps> = (
         'limel-select': true,
         'mdc-select': true,
         'mdc-select--disabled': props.disabled,
+        'limel-select--readonly': props.readonly,
         'limel-select--required': props.required,
         'limel-select--invalid': !isValid,
         'limel-select--empty': !hasValue,
@@ -53,7 +55,7 @@ export const SelectTemplate: FunctionalComponent<SelectTemplateProps> = (
 
     return (
         <div class={classList}>
-            <SelectValue {...props} hasValue={hasValue} />
+            <SelectValue {...props} hasValue={hasValue} isValid={isValid} />
             <HelperText text={props.helperText} />
             <SelectDropdown {...props} />
         </div>
@@ -63,6 +65,7 @@ export const SelectTemplate: FunctionalComponent<SelectTemplateProps> = (
 const SelectValue: FunctionalComponent<
     SelectTemplateProps & {
         hasValue: boolean;
+        isValid: boolean;
     }
 > = (props) => {
     const containerClassList = {
@@ -88,9 +91,24 @@ const SelectValue: FunctionalComponent<
                     {getSelectedText(props.value)}
                 </span>
             </div>
+            <ShowIcon {...props} isValid={props.isValid} />
             <i class="mdc-select__dropdown-icon" />
             <label class={labelClassList}>{props.label}</label>
         </div>
+    );
+};
+
+const ShowIcon: FunctionalComponent<
+    SelectTemplateProps & {
+        isValid: boolean;
+    }
+> = (props) => {
+    if (props.isValid) {
+        return;
+    }
+
+    return (
+        <limel-icon name="high_importance" size="medium" class="invalid-icon" />
     );
 };